video-processor/requirements.txt
praw==7.7.1
yt-dlp==2024.03.10
moviepy==1.0.3
gTTS==2.5.1
openai-whisper==20231117
edge-tts==6.1.10
alt-profanity-check==1.3.0
python-dotenv==0.21.1
torch==2.3.1
torchaudio==2.3.1
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0 
requests
beautifulsoup4
alt-profanity-check 